+++ CHANGELOG.md
@@ -32,6 +32,9 @@ This beta release is built on top of [OpenTelemetry .NET](https://github.com/ope
 
 ### Fixed
 
+- Fix console error messages `Log: Exception creating FileSink`
+ [#1885](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/issues/1885)
+
 ### Security
 
 ## [0.5.1-beta.3](https://github.com/open-telemetry/opentelemetry-dotnet-instrumentation/releases/tag/v0.5.1-beta.3)